Join facets components in a single one into the playground
