+++ packages/_vue3-migration-test/src/main.ts
@@ -2,9 +2,10 @@ import { XComponentsAdapter } from '@empathyco/x-types';
 import { Component, configureCompat, createApp } from 'vue';
 import { createStore } from 'vuex';
 import { xPlugin } from '../../x-components/src/plugins/x-plugin';
+import { facetsXModule } from '../../x-components/src/x-modules/facets/x-module';
 import App from './App.vue';
 import router from './router';
-import { facetsXModule, scrollXModule } from './';
+import { scrollXModule, testFacetsXModule } from './';
 
 // Warnings that cannot be solved in Vue 2 (a.k.a. breaking  changes) are suppressed
 const VUE_COMPAT_MODE = Number(import.meta.env.VITE_VUE_COMPAT_MODE);
@@ -34,8 +35,9 @@ createApp(App as Component)
   .use(xPlugin, {
     adapter,
     store,
+    initialXModules: [facetsXModule],
     __PRIVATE__xModules: {
-      facets: facetsXModule,
+      facets: testFacetsXModule,
       scroll: scrollXModule
     }
   })
