--- packages/_vue3-migration-test/src/x-modules/facets/components/facets/index.ts
@@ -1,1 +1,0 @@
-export { default as TestFacets } from './test-facets.vue';
