+++ packages/_vue3-migration-test/src/x-modules/facets/components/index.ts
@@ -1,3 +1,1 @@
-export * from './facets';
-export * from './lists';
-export { default as TestClearFilters } from './test-clear-filters.vue';
+export { default as TestFacets } from './test-facets.vue';
