--- packages/_vue3-migration-test/src/x-modules/facets/components/lists/index.ts
@@ -1,2 +1,0 @@
-export { default as TestSelectedFilters } from './test-selected-filters.vue';
-export { default as TestSelectedFiltersList } from './test-selected-filters-list.vue';
