--- packages/_vue3-migration-test/src/x-modules/facets/components/lists/test-selected-filters-list.vue
@@ -1,9 +1,0 @@
-<template>
-  <SelectedFiltersList :facetsIds="['gender', 'brand']" alwaysVisible>
-    <template #default="{ filter }">{{ filter.facetId }}: {{ filter.label }}</template>
-  </SelectedFiltersList>
-</template>
-
-<script setup lang="ts">
-  import SelectedFiltersList from '../../../../../../x-components/src/x-modules/facets/components/lists/selected-filters-list.vue';
-</script>
