--- packages/_vue3-migration-test/src/x-modules/facets/components/lists/test-selected-filters.vue
@@ -1,11 +1,0 @@
-<template>
-  <SelectedFilters :facetsIds="['gender', 'brand']" alwaysVisible>
-    <template #default="{ selectedFilters }">
-      Selected filters: {{ selectedFilters.length }}
-    </template>
-  </SelectedFilters>
-</template>
-
-<script setup lang="ts">
-  import SelectedFilters from '../../../../../../x-components/src/x-modules/facets/components/lists/selected-filters.vue';
-</script>
