--- packages/_vue3-migration-test/src/x-modules/facets/components/test-clear-filters.vue
@@ -1,11 +1,0 @@
-<template>
-  <ClearFilters :facetsIds="['gender', 'brand']" alwaysVisible>
-    <template #default="{ selectedFilters }">
-      Clear filters ({{ selectedFilters.length }} selected)
-    </template>
-  </ClearFilters>
-</template>
-
-<script setup lang="ts">
-  import ClearFilters from '../../../../../x-components/src/x-modules/facets/components/clear-filters.vue';
-</script>
