+++ packages/_vue3-migration-test/src/x-modules/facets/components/test-facets.vue
@@ -1,0 +1,40 @@
+<template>
+  <div>
+    <h2>Facets</h2>
+    <Facets :renderableFacets="facetsIds.join(',')">
+      <template #default="{ facet }">
+        <h3>{{ facet.label }}</h3>
+        <ul>
+          <li v-for="filter in facet.filters" :key="filter.id">
+            {{ filter.label }}
+          </li>
+        </ul>
+      </template>
+    </Facets>
+    <h2>SelectedFilters</h2>
+    <SelectedFilters :facetsIds="facetsIds" alwaysVisible>
+      <template #default="{ selectedFilters }">
+        Selected filters: {{ selectedFilters.length }}
+      </template>
+    </SelectedFilters>
+    <h2>SelectedFiltersList</h2>
+    <SelectedFiltersList :facetsIds="facetsIds" alwaysVisible>
+      <template #default="{ filter }">{{ filter.facetId }}: {{ filter.label }}</template>
+    </SelectedFiltersList>
+    <h2>ClearFilters</h2>
+    <ClearFilters :facetsIds="facetsIds" alwaysVisible>
+      <template #default="{ selectedFilters }">
+        Clear filters ({{ selectedFilters.length }} selected)
+      </template>
+    </ClearFilters>
+  </div>
+</template>
+
+<script setup lang="ts">
+  import ClearFilters from '../../../../../x-components/src/x-modules/facets/components/clear-filters.vue';
+  import Facets from '../../../../../x-components/src/x-modules/facets/components/facets/facets.vue';
+  import SelectedFilters from '../../../../../x-components/src/x-modules/facets/components/lists/selected-filters.vue';
+  import SelectedFiltersList from '../../../../../x-components/src/x-modules/facets/components/lists/selected-filters-list.vue';
+
+  const facetsIds = ['gender', 'brand'];
+</script>
